Restructured code and added DGSEM discretization
The name of the repository is changed to
Rotating_Shallow_Water_Verification_Suite. The code is thoroughly modified,
restructured, and organized into the following directories:
(a) src containing the subdirectories
    (i)  MPAS_Ocean_Shallow_Water
    (ii) DGSEM_Rotating_Shallow_Water
(b) tests containing the subdirectories
    (i)  MPAS_Ocean_Shallow_Water_Tests
    (ii) DGSEM_Rotating_Shallow_Water_Tests
(c) tutorials containing the subdirectories
    (i)  MPAS_Ocean_Shallow_Water_Tutorials
    (ii) DGSEM_Rotating_Shallow_Water_Tutorials
(d) meshes containing the subdirectory
    (i)  MPAS_Ocean_Shallow_Water_Meshes
(e) output containing the subdirectories
    (i)  MPAS_Ocean_Shallow_Water_Output
    (ii) DGSEM_Rotating_Shallow_Water_Output
As the names imply, the src directory contains the source code, the tests
directory contains code for testing the functions in the scripts of the src
directory. The scripts in the tutorials directory are exact replicas of the ones
in the tests directory but in Jupiter Notebook format. It is up to the user to
choose between a specific Python script in the tests directory and the
equivalent Jupyter Notebook in the tutorials directory to test one or more
functions in the corresponding script in the src directory. The reason the
Jupyter Notebooks are included is because it is sometimes convenient to run only
a few cells of the Jupiter Notebook, and check out any error message or the
output of print statements right below the cells as opposed to all over the
terminal. The code is made significantly more modular and the number of scripts
are increased manyfold. It is impossible to list every modification in this
commit within the span of a few sentences or even a few paragraphs. So, we list
the major modification below, and refer the interested user to the the header
sections of the Python scripts for further information on their respective
roles:
(a) Each of the src, tests, tutorials, and output directories contains two
    subdirectories corresponding to two types of spatial discretizations:
    (i)  the already existing TRiSK-based mimetic finite volume method used in
         MPAS-Ocean; and
    (ii) a newly added discontinuous Galerkin Spectral Element Method.
(b) The README file is modified to reflect the changes to the repository.
(c) The MPAS_Ocean_Shallow_Water_Meshes subdirectory within the meshes directory
    should contain the planar hexagonal MPAS-Ocean mesh files for simulating the
    numerical solutions and conducting convergence studies. These mesh files
    being of considerable size, are not included within the Github repository.
    But the download link for the entire MPAS_Ocean_Shallow_Water_Meshes
    subdirectory is provided in the new README file.
(d) The output files generated by the two types of spatial discretizations will
    be placed in the respective subdirectories within the output directory.
(e) The derivation of the state variables and source terms of the planetary and
    topographic Rossby waves using the symbolic algebra package SymPy is
    removed. This is because the previous formulation of the Rossby waves as a
    manufactured solution results in an ill-posed system of inhomogeneous
    shallow water equations that does not conserve the total energy and renders
    the numerical solution unstable. The new formulation of the Rossby waves,
    which are implemented only for the DGSEM spatial discretization,
    (i)   are not equipped with exact solutions;
    (ii)  involve the propagation of an initially geostrophic eddy on a beta
          plane (for the planetary Rossby wave) or a sloping bottom (for the
          topographic Rossby wave); and
    (iii) result in stable numerical solutions.
(f) Some important functions are removed only to be improved and put back in a
    later commit. These include functions for plotting the exact solution along
    a relevant section of the domain, and checking the order of convergence of
    the spatial operators of the TRiSK-based mimetic finite volume method.
(g) The diffusion equation and the viscous Burgers equation test cases are
    removed.
(h) A plane Gaussian wave test case is added for testing the DGSEM code.
(i) A non-linear manufactured solution test case is added for both types of
    spatial discretizations. It tests the numerical implementation of the
    non-linear advection terms in the momentum and continuity equations.
(j) The numerical solutions are tested for accuracy against their exact
    counterparts (except for the Rossby waves) as well as for stability.
    Convergence studies are conducted with refinement in both space and time
    (except for the Rossby waves), only in space, and only in time.
(k) Since the layer-averaged barotropic variables are two-dimensional, the array
    dimension representing the layer index of these variables is removed. As a
    result, at a given point in time, these variables are now represented by
    one-dimensional arrays with the (only) array index representing a cell
    center or an edge center or a node depending on the locations of these
    variables.
(l) The script Filter_Routines.py is removed, since the barotropic-baroclinic
    splitting and the time-averaging of the barotropic modes using a filter do
    not apply to the shallow water system of equations involving a single layer
    in the vertical direction.
(m) A license file is included.
(n) The makefile is removed.