Modified spacings, plotting, and compacted lines
(a) Adjusted spacings and compacted a few lines of code.

(b) In the convergence studies of Rossby waves only in time, employing TRiSK and
    fourth-order Runge-Kutta and Adams-Bashforth time integrators, it was
    observed that machine precision error dominates discretization error,
    leading to converge slopes deviating from theoretical predictions.
    Consequently, these results have been excluded from the plotting process.

(c) In the convergence plots of Rossby waves only in space using TRiSK, the
    line-styles of the convergence curves representing the different time
    integrators have been distinctly varied to ensure clear differentiation when
    superimposed.

(d) Updated the meshes/README.md file to include a reference to the most recent
    Zenodo record for the MPAS_Ocean_Shallow_Water_Meshes.

(e) Updated Zenodo DOI in README.md and CITATION.cff files.